third_party/WebKit/LayoutTests/bluetooth/service/device-same-object.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return getHealthThermometerDevice({
      filters: [{services: ['health_thermometer']}]})
    .then(({device}) => device.gatt.getPrimaryService('health_thermometer'))
    .then(service => {
      assert_equals(service.device, service.device);
    });
}, "[SameObject] test for BluetoothRemoteGATTService's device.");
</script>
